cleanup: remove orphan command modules
