+++ cli/internal/cmd/browser.go
@@ -1,0 +1,18 @@
+package cmd
+
+import (
+	"os/exec"
+	"runtime"
+)
+
+// openBrowser opens a URL in the default browser (cross-platform).
+func openBrowser(url string) error {
+	switch runtime.GOOS {
+	case "darwin":
+		return exec.Command("open", url).Start()
+	case "windows":
+		return exec.Command("cmd", "/c", "start", url).Start()
+	default:
+		return exec.Command("xdg-open", url).Start()
+	}
+}
